agent-connection/requirements.txt
# Reuters RUM Sales Intelligence Agent – Python dependencies
snowflake-connector-python>=3.0.0
anthropic>=0.25.0
tabulate>=0.9.0
python-dotenv>=1.0.0
